refactor: move mixin property service out of bootstrapper perf: use strings as keys
should help with memory usage, as we no longer have to keep around references to keys
