+++ plugin-mixin-bootstrapper/src/main/resources/META-INF/services/org.spongepowered.asm.service.IGlobalPropertyService
@@ -1,1 +1,1 @@
-pw.kaboom.papermixins.pluginmixin.bootstrap.Bootstrapper
+pw.kaboom.papermixins.pluginmixin.bootstrap.SimpleMixinPropertyService
